Namespace properly, fix bug with callbacks
- to distinguish which WhatsApp integration the process webhooks job is
  for.
- the callbacks were left in the webhook controller/concerns and not
  moved to the process webhook job
